+++ docs/conf.py
@@ -9,7 +9,7 @@
 # If extensions (or modules to document with autodoc) are in another directory,
 # add these directories to sys.path here. If the directory is relative to the
 # documentation root, use os.path.abspath to make it absolute, like shown here.
-import os
+import os.path
 import sys
 
 sys.path.insert(0, os.path.abspath(".."))